+++ CHANGES.rst
@@ -4,7 +4,13 @@ Changelog
 1.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Allow all Managers to use the full control panel.
+  Until now, for some parts you needed to be Manager in the Zope root.
+  Fixes `issue 30 <https://github.com/collective/collective.revisionmanager/issues/30>`_.
+  [maurits]
+
+- Test with Plone 6 (still also on 4.3, 5.1, 5,2) and on GitHub Actions.
+  [maurits]
 
 
 1.1.0 (2021-09-16)